Renamed ee10 client transport tests package.
