+++ jetty-ee10/jetty-ee10-tests/jetty-ee10-test-client-transports/src/test/java/org/eclipse/jetty/ee10/test/client/transport/ServerTimeoutsTest.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.ee10.http.client;
+package org.eclipse.jetty.ee10.test.client.transport;
 
 import java.io.IOException;
 import java.io.InterruptedIOException;
@@ -52,9 +52,9 @@ import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
-import static org.eclipse.jetty.ee10.http.client.Transport.FCGI;
-import static org.eclipse.jetty.ee10.http.client.Transport.H2;
-import static org.eclipse.jetty.ee10.http.client.Transport.H2C;
+import static org.eclipse.jetty.ee10.test.client.transport.Transport.FCGI;
+import static org.eclipse.jetty.ee10.test.client.transport.Transport.H2;
+import static org.eclipse.jetty.ee10.test.client.transport.Transport.H2C;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
